backend/requirements.txt
asgiref==3.10.0
channels==4.1.0
channels-redis==4.2.0
cryptography==44.0.0
daphne==4.1.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==24.3
django-import-export==4.3.3
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.0
openpyxl==3.1.5
pandas==2.2.2
pillow==12.0.0
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
requests==2.32.5

